+++ cognite/client/_api/assets.py
@@ -5,7 +5,15 @@ from typing import *
 
 from cognite.client import utils
 from cognite.client._api_client import APIClient
-from cognite.client.data_classes import Asset, AssetAggregate, AssetFilter, AssetList, AssetUpdate, TimestampRange
+from cognite.client.data_classes import (
+    Asset,
+    AssetAggregate,
+    AssetFilter,
+    AssetList,
+    AssetUpdate,
+    TimestampRange,
+    LabelFilter,
+)
 from cognite.client.exceptions import CogniteAPIError
 
 
@@ -26,7 +34,7 @@ class AssetsAPI(APIClient):
         metadata: Dict[str, str] = None,
         data_set_ids: List[int] = None,
         data_set_external_ids: List[str] = None,
-        labels: List = None,
+        labels: LabelFilter = None,
         source: str = None,
         created_time: Union[Dict[str, Any], TimestampRange] = None,
         last_updated_time: Union[Dict[str, Any], TimestampRange] = None,
@@ -51,7 +59,7 @@ class AssetsAPI(APIClient):
             metadata (Dict[str, str]): Custom, application specific metadata. String key -> String value
             data_set_ids (List[int]): Return only assets in the specified data sets with these ids.
             data_set_external_ids (List[str]): Return only assets in the specified data sets with these external ids.
-            labels (List): List of label filters
+            labels (LabelFilter): Return only the assets matching the specified label.
             source (str): The source of this asset
             created_time (Union[Dict[str, int], TimestampRange]):  Range between two timestamps. Possible keys are `min` and `max`, with values given as time stamps in ms.
             last_updated_time (Union[Dict[str, int], TimestampRange]):  Range between two timestamps. Possible keys are `min` and `max`, with values given as time stamps in ms.
@@ -180,7 +188,7 @@ class AssetsAPI(APIClient):
         asset_subtree_external_ids: List[str] = None,
         data_set_ids: List[int] = None,
         data_set_external_ids: List[str] = None,
-        labels: List = None,
+        labels: LabelFilter = None,
         metadata: Dict[str, str] = None,
         source: str = None,
         created_time: Union[Dict[str, Any], TimestampRange] = None,
@@ -203,7 +211,7 @@ class AssetsAPI(APIClient):
             asset_subtree_external_ids (List[str]): List of asset subtrees external ids to filter on.
             data_set_ids (List[int]): Return only assets in the specified data sets with these ids.
             data_set_external_ids (List[str]): Return only assets in the specified data sets with these external ids.
-            labels (List): List of label filters.
+            labels (LabelFilter): Return only the assets matching the specified label filter.
             metadata (Dict[str, str]): Custom, application specific metadata. String key -> String value.
             source (str): The source of this asset.
             created_time (Union[Dict[str, int], TimestampRange]):  Range between two timestamps. Possible keys are `min` and `max`, with values given as time stamps in ms.
@@ -239,6 +247,14 @@ class AssetsAPI(APIClient):
                 >>> c = CogniteClient()
                 >>> for asset_list in c.assets(chunk_size=2500):
                 ...     asset_list # do something with the assets
+
+            Filter assets based on labels::
+
+                >>> from cognite.client import CogniteClient
+                >>> from cognite.client.data_classes import LabelFilter
+                >>> c = CogniteClient()
+                >>> my_label_filter = LabelFilter(contains_all=["PUMP", "VERIFIED"])
+                >>> asset_list = c.assets.list(labels=my_label_filter)
         """
         if aggregated_properties:
             aggregated_properties = [utils._auxiliary.to_camel_case(s) for s in aggregated_properties]
@@ -313,6 +329,14 @@ class AssetsAPI(APIClient):
                 >>> from cognite.client.data_classes import Asset
                 >>> c = CogniteClient()
                 >>> assets = [Asset(name="asset1"), Asset(name="asset2")]
+                >>> res = c.assets.create(assets)
+
+            Create asset with label::
+
+                >>> from cognite.client import CogniteClient
+                >>> from cognite.client.data_classes import Asset, Label
+                >>> c = CogniteClient()
+                >>> asset = Asset(name="my_pump", labels=[Label(external_id="PUMP")])
                 >>> res = c.assets.create(assets)
         """
         utils._auxiliary.assert_type(asset, "asset", [Asset, list])
@@ -404,12 +428,20 @@ class AssetsAPI(APIClient):
                 >>> my_update = AssetUpdate(id=1).description.set("New description").metadata.add({"key": "value"})
                 >>> res = c.assets.update(my_update)
 
-            Perform a partial update on a asset, attach a label to an asset::
+            Attach a label to an asset::
 
                 >>> from cognite.client import CogniteClient
                 >>> from cognite.client.data_classes import AssetUpdate
                 >>> c = CogniteClient()
                 >>> my_update = AssetUpdate(id=1).labels.add(["PUMP"])
+                >>> res = c.assets.update(my_update)
+
+            Detach a label from an asset::
+
+                >>> from cognite.client import CogniteClient
+                >>> from cognite.client.data_classes import AssetUpdate
+                >>> c = CogniteClient()
+                >>> my_update = AssetUpdate(id=1).labels.remove(["PUMP"])
                 >>> res = c.assets.update(my_update)
         """
         return self._update_multiple(items=item)